Add files via upload
A game in which you input a range of numbers minimum number, maximum number, and a common divisor of all these numbers. Then the computer picks a random number from among this range, and you attempt to guess what that number is. If your guess is lower, it will tell you your guess is too low, and if its higher, it will tell you the guess is too high.
Zarik Khan - 11/29/2020